+++ config/packages/dev/old_sound_rabbit_mq.yaml
@@ -2,13 +2,6 @@ services:
     _defaults:
         public: true
 
-#    old_sound_rabbit_mq.content_push_producer:
-#        class: SWP\Bundle\CoreBundle\Tests\Producer\TestProducer
-#        arguments:
-#            - '@SWP\Bundle\CoreBundle\Consumer\ContentPushConsumer'
-#    old_sound_rabbit_mq.migration_content_push_producer:
-#        alias: old_sound_rabbit_mq.content_push_producer
-
     old_sound_rabbit_mq.image_conversion_producer:
         class: SWP\Bundle\CoreBundle\Tests\Producer\TestProducer
         arguments:
