+++ etc/scripts/supervisor/messenger-worker.conf
@@ -1,6 +1,6 @@
 [program:messenger-consume]
 directory=/var/www/publisher/
-command=php bin/console messenger:consume async_content_push async_analytics_event async_webhooks async_analytics_export --time-limit=3600
+command=php bin/console messenger:consume async_content_push async_analytics_event async_webhooks async_analytics_export async_migration_content_push --time-limit=3600
 numprocs=2
 autostart=true
 autorestart=true
