+++ src/SWP/Bundle/CoreBundle/Resources/config/app/config.yml
@@ -325,20 +325,12 @@ old_sound_rabbit_mq:
         image_conversion:
             connection:       default
             exchange_options: {name: '%kernel.environment%_%kernel.secret%_image_conversion', type: direct}
-        migration_content_push:
-            connection:       default
-            exchange_options: {name: '%kernel.environment%_%kernel.secret%_migration_content_push', type: direct}
     consumers:
         image_conversion:
             connection:       default
             exchange_options: {name: '%kernel.environment%_%kernel.secret%_image_conversion', type: direct}
             queue_options:    {name: '%kernel.environment%_%kernel.secret%_image_conversion'}
             callback:         SWP\Bundle\CoreBundle\Consumer\ImageConversionConsumer
-#        migration_content_push:
-#            connection:       default
-#            exchange_options: {name: '%kernel.environment%_%kernel.secret%_migration_content_push', type: direct}
-#            queue_options:    {name: '%kernel.environment%_%kernel.secret%_migration_content_push'}
-#            callback:         SWP\Bundle\CoreBundle\Consumer\ContentPushConsumer
 
 bazinga_hateoas:
     twig_extension:
